+++ CHANGELOG.md
@@ -15,3 +15,13 @@ the following form: YYYY.0M.0D.
 - Designed a PriorityQueue component
 - Designed a SkillTree component
 - Designed a PrefixTree component
+
+## [2025.10.17]
+
+### Added
+
+- Designed a proof of concept for SkillTree component
+
+### Updated
+
+- Changed design to include dependency-based skill unlocking, depth calculation, and available skill detection methods.